Move OptionalGuidAssertions to FluentAssertions.Optional.Primitives namespace
